Move dash data to private api router
