+++ src/test/server/apis/incoming-prs-test.ts
@@ -1,10 +1,11 @@
 import anyTest, {TestInterface} from 'ava';
 
-import {fetchIncomingData} from '../../../server/apis/dash-data';
+import {handleIncomingPRRequest} from '../../../server/apis/dash-data/handle-incoming-pr-request';
 import {IncomingDashResponse, PullRequest} from '../../../types/api';
 import {PullRequestReviewState} from '../../../types/gql-types';
 import {initFirestore} from '../../../utils/firestore';
 import {initGithub} from '../../../utils/github';
+import {newFakeRequest} from '../../utils/newFakeRequest';
 import {newFakeUserRecord} from '../../utils/newFakeUserRecord';
 import {startTestReplayServer} from '../../utils/replay-server';
 
@@ -27,17 +28,18 @@ test.beforeEach(async (t) => {
   initGithub(url, url);
 
   const userRecord = newFakeUserRecord();
-  const data = await fetchIncomingData(userRecord, 'project-health1');
+  userRecord.username = 'project-health1';
+  const response = await handleIncomingPRRequest(newFakeRequest(), userRecord);
   server.close();
 
   const prsById = new Map();
-  for (const pr of data.prs) {
+  for (const pr of response.data.prs) {
     prsById.set(
         pr.url.replace('https://github.com/project-health1/repo/pull/', ''),
         pr);
   }
   t.context = {
-    data,
+    data: response.data,
     prsById,
   };
 });
